Move `StandardKeyboard` inside `Keyboard` object
