+++ lib/profanity/src/core/profanity.Keyboard.scala
@@ -32,7 +32,16 @@
                                                                                                   */
 package profanity
 
+import anticipation.*
+import contingency.*
+import distillate.*
+import gossamer.*
+import hypotenuse.*
+import parasite.*
 import proscenium.*
+import rudiments.*
+import spectacular.*
+import vacuous.*
 
 trait Keyboard:
   type Keypress
@@ -69,3 +78,77 @@ object Keyboard:
     case '5'       => Keypress.PageUp
     case '6'       => Keypress.PageDown
     case _         => Keypress.Escape
+
+  class Standard()(using Monitor, Codicil) extends Keyboard:
+    type Keypress = profanity.Keypress | TerminalInfo
+
+    def process(stream: Stream[Char]): Stream[Keypress] = stream match
+      case '\u001b' #:: rest =>
+        safely(async(rest.head).await(30L)) match
+          case Unset => Keypress.Escape #:: process(rest)
+          case _ => rest match
+            case 'O' #:: key #:: rest => Keypress.FunctionKey(key.toInt - 79) #:: process(rest)
+            case '[' #:: rest        => rest match
+              case (code@('A' | 'B' | 'C' | 'D' | 'F' | 'H')) #:: rest =>
+                Keyboard.navigation(code) #:: process(rest)
+
+              case code #:: '~' #:: rest if '1' <= code <= '9' =>
+                Keyboard.vt(code) #:: process(rest)
+
+              case code #:: ';' #:: modifiers #:: '~' #:: rest if '1' <= code <= '9' =>
+                Keyboard.modified(modifiers, Keyboard.vt(code)) #:: process(rest)
+
+              case '1' #:: ';' #:: modifiers #:: (code@('A' | 'B' | 'C' | 'D' | 'F' | 'H'))
+                  #:: rest =>
+                Keyboard.modified(modifiers, Keyboard.navigation(code)) #:: process(rest)
+
+              case '2' #:: '0' #:: '0' #:: '~' #:: tail =>
+                val size = tail.indexOfSlice(List('\u001b', '[', '2', '0', '1', '~'))
+                val content = tail.take(size).map(_.show).join
+                TerminalInfo.Paste(content) #:: process(tail.drop(size + 6))
+
+              case other =>
+                val sequence = other.takeWhile(!_.isLetter)
+
+                other.drop(sequence.length) match
+                  case 'R' #:: tail => sequence.map(_.show).join.cut(';').to(List) match
+                    case List(As[Int](rows), As[Int](cols)) =>
+                      TerminalInfo.WindowSize(rows, cols) #:: process(tail)
+
+                    case _ =>
+                      TerminalInfo.WindowSize(20, 30) #:: process(tail)
+
+                  case 'O' #:: tail =>
+                    TerminalInfo.LoseFocus #:: process(tail)
+
+                  case 'I' #:: tail =>
+                    TerminalInfo.GainFocus #:: process(tail)
+
+                  case char #:: tail =>
+                    Keypress.EscapeSeq(char, sequence*) #:: process(tail)
+
+                  case _ =>
+                    Stream()
+
+            case ']' #:: '1' #:: '1' #:: ';' #:: 'r' #:: 'g' #:: 'b' #:: ':' #:: rest =>
+              val content = rest.takeWhile(_ != '\u001b').mkString.tt
+              val continuation = rest.drop(content.length + 2)
+
+              content.cut(t"/").to(List) match
+                case List(red, green, blue) =>
+                  def decimal(hex: Text): Int = Integer.parseInt(hex.s, 16)
+                  TerminalInfo.BgColor(decimal(red), decimal(green), decimal(blue))
+                  #:: process(continuation)
+
+                case _ =>
+                  process(continuation)
+
+            case rest =>
+              process(rest)
+
+      case ('\b' | '\u007f') #:: rest     => Keypress.Backspace #:: process(rest)
+      case '\u0009' #:: rest              => Keypress.Tab #:: process(rest)
+      case ('\u000a' | '\u000d') #:: rest => Keypress.Enter #:: process(rest)
+      case CtrlChar(char) #:: rest        => Keypress.Ctrl(char) #:: process(rest)
+      case other #:: rest                 => Keypress.CharKey(other) #:: process(rest)
+      case _                              => Stream()
