+++ lib/profanity/src/core/profanity.Terminal.scala
@@ -55,7 +55,7 @@ extends Interactivity[TerminalEvent]:
 
   export context.stdio.{in, out, err}
 
-  val keyboard: StandardKeyboard = StandardKeyboard()
+  val keyboard: Keyboard.Standard = Keyboard.Standard()
   val rows0: Promise[Int] = Promise()
   val columns0: Promise[Int] = Promise()
   var mode: Optional[Luminance] = Unset
